Move ClaimingForListBuilder to controllers.helpers package
